主要是一些小修小补，以及test路径问题的修复 feat(configs): update training configs and refactor scripts
- Change default device from CPU to CUDA in base.yaml
- Reduce batch size from 32 to 16 in training.yaml
- Disable ElasticFace loss and set type to "cos+" in training.yaml
- Update model from yolov8n-cls.pt to yolo26x-cls.pt
- Add ElasticFace type parameter support to training scripts
- Refactor auto-annotation tests to use temporary paths
- Move VLM test script from scripts to tools directory
- Add package initialization for training scripts
- Optimize HDBSCAN outlier detection with np.flatnonzero
- Remove old test files and add new YOLO model files
