+++ training/scripts/__init__.py
@@ -1,0 +1,1 @@
+"""Training scripts package."""
