+++ training/tests/scripts/test_elasticface_loss.py
@@ -479,6 +479,209 @@ class TestElasticFaceParameters:
         assert loss_fn.std == 0.01
 
 
+class TestElasticFaceTypes:
+    """Test ElasticFace loss types: cos, arc, cos+, arc+."""
+
+    @pytest.fixture
+    def sample_embeddings(self):
+        """Create sample embedding tensors for testing."""
+        embeddings = torch.randn(8, 256)
+        return torch.nn.functional.normalize(embeddings, p=2, dim=1)
+
+    @pytest.fixture
+    def sample_labels(self):
+        """Create sample labels for testing."""
+        return torch.randint(0, 10, (8,))
+
+    def test_elastic_face_type_cos(self, sample_embeddings, sample_labels):
+        """Test ElasticFace with type='cos' (CosFace)."""
+        from training_utils import ElasticFaceArcFace
+
+        loss_fn = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.01,
+            type="cos",
+        )
+
+        logits = loss_fn(sample_embeddings, sample_labels)
+
+        assert logits.shape == (8, 10), f"Expected shape (8, 10), got {logits.shape}"
+        assert loss_fn.type == "cos", "Type should be 'cos'"
+
+    def test_elastic_face_type_arc(self, sample_embeddings, sample_labels):
+        """Test ElasticFace with type='arc' (ArcFace)."""
+        from training_utils import ElasticFaceArcFace
+
+        loss_fn = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.01,
+            type="arc",
+        )
+
+        logits = loss_fn(sample_embeddings, sample_labels)
+
+        assert logits.shape == (8, 10), f"Expected shape (8, 10), got {logits.shape}"
+        assert loss_fn.type == "arc", "Type should be 'arc'"
+
+    def test_elastic_face_type_cos_plus(self, sample_embeddings, sample_labels):
+        """Test ElasticFace with type='cos+' (CosFace+ adaptive)."""
+        from training_utils import ElasticFaceArcFace
+
+        loss_fn = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.01,
+            type="cos+",
+        )
+
+        logits = loss_fn(sample_embeddings, sample_labels)
+
+        assert logits.shape == (8, 10), f"Expected shape (8, 10), got {logits.shape}"
+        assert loss_fn.type == "cos+", "Type should be 'cos+'"
+        assert loss_fn.plus is True, "plus should be True for 'cos+'"
+
+    def test_elastic_face_type_arc_plus(self, sample_embeddings, sample_labels):
+        """Test ElasticFace with type='arc+' (ArcFace+ adaptive)."""
+        from training_utils import ElasticFaceArcFace
+
+        loss_fn = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.01,
+            type="arc+",
+        )
+
+        logits = loss_fn(sample_embeddings, sample_labels)
+
+        assert logits.shape == (8, 10), f"Expected shape (8, 10), got {logits.shape}"
+        assert loss_fn.type == "arc+", "Type should be 'arc+'"
+        assert loss_fn.plus is True, "plus should be True for 'arc+'"
+
+    def test_elastic_face_all_types_produce_valid_logits(
+        self, sample_embeddings, sample_labels
+    ):
+        """Test that all types produce valid logits."""
+        from training_utils import ElasticFaceArcFace
+
+        for loss_type in ["cos", "arc", "cos+", "arc+"]:
+            loss_fn = ElasticFaceArcFace(
+                in_features=256,
+                out_features=10,
+                s=30.0,
+                m=0.30,
+                std=0.01,
+                type=loss_type,
+            )
+
+            logits = loss_fn(sample_embeddings, sample_labels)
+
+            # Check shape
+            assert logits.shape == (8, 10), f"Shape mismatch for type='{loss_type}'"
+
+            # Check no NaN
+            assert not torch.isnan(logits).any(), f"NaN detected for type='{loss_type}'"
+
+            # Check finite values
+            assert torch.isfinite(logits).all(), f"Non-finite values for type='{loss_type}'"
+
+    def test_elastic_face_type_cos_vs_arc_difference(
+        self, sample_embeddings, sample_labels
+    ):
+        """Test that cos and arc types produce different results."""
+        from training_utils import ElasticFaceArcFace
+
+        # Set seed for reproducibility
+        torch.manual_seed(42)
+
+        loss_fn_cos = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.0,  # Fixed margin for comparison
+            type="cos",
+        )
+
+        torch.manual_seed(42)
+
+        loss_fn_arc = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+            s=30.0,
+            m=0.30,
+            std=0.0,  # Fixed margin for comparison
+            type="arc",
+        )
+
+        # Use the same kernel weights for fair comparison
+        loss_fn_arc.kernel.data.copy_(loss_fn_cos.kernel.data)
+
+        logits_cos = loss_fn_cos(sample_embeddings, sample_labels)
+        logits_arc = loss_fn_arc(sample_embeddings, sample_labels)
+
+        # Results should be different (cos applies margin in cosine space,
+        # arc applies margin in angular space)
+        assert not torch.allclose(
+            logits_cos, logits_arc, rtol=1e-3, atol=1e-5
+        ), "cos and arc should produce different results"
+
+    def test_elastic_face_wrapper_with_type(self, sample_embeddings, sample_labels):
+        """Test ElasticFaceLossWrapper with different types."""
+        from training_utils import ElasticFaceLossWrapper, l2_norm
+
+        for loss_type in ["cos", "arc", "cos+", "arc+"]:
+            wrapper = ElasticFaceLossWrapper(
+                num_classes=10,
+                embedding_dim=256,
+                lambda_weight=0.5,
+                s=30.0,
+                m=0.30,
+                std=0.01,
+                type=loss_type,
+            )
+
+            normalized_embeddings = l2_norm(sample_embeddings, axis=1)
+            total_loss, loss_dict = wrapper(normalized_embeddings, sample_labels)
+
+            assert isinstance(total_loss, torch.Tensor), f"Loss should be tensor for type='{loss_type}'"
+            assert total_loss.item() > 0, f"Loss should be positive for type='{loss_type}'"
+            assert not torch.isnan(total_loss), f"Loss should not be NaN for type='{loss_type}'"
+            assert wrapper.elastic_face.type == loss_type, f"Type mismatch for type='{loss_type}'"
+
+    def test_elastic_face_default_type_is_arc(self):
+        """Test that default type is 'arc' for backward compatibility."""
+        from training_utils import ElasticFaceArcFace
+
+        # Test that default type is 'arc' when not specified
+        loss_fn_default = ElasticFaceArcFace(
+            in_features=256,
+            out_features=10,
+        )
+
+        assert loss_fn_default.type == "arc", "Default type should be 'arc'"
+
+    def test_elastic_face_invalid_type_raises_error(self):
+        """Test that invalid type raises ValueError."""
+        from training_utils import ElasticFaceArcFace
+
+        with pytest.raises(ValueError, match="Invalid type"):
+            ElasticFaceArcFace(
+                in_features=256,
+                out_features=10,
+                type="invalid_type",
+            )
+
+
 class TestTrainClassifyIntegration:
     """Test ElasticFace integration in train_classify.py."""
 
@@ -500,7 +703,7 @@ class TestTrainClassifyIntegration:
         assert "--elastic-s" in result.stdout
         assert "--elastic-m" in result.stdout
         assert "--elastic-std" in result.stdout
-        assert "--elastic-plus" in result.stdout
+        assert "--elastic-type" in result.stdout
 
     def test_import_training_utils_elasticface(self):
         """Test that training_utils can be imported with ElasticFace classes."""
@@ -540,14 +743,20 @@ class TestTrainClassifyIntegration:
         m = config.get("training.advanced.elastic_face.m")
         std = config.get("training.advanced.elastic_face.std")
         plus = config.get("training.advanced.elastic_face.plus")
+        elastic_type = config.get("training.advanced.elastic_face.type")
 
-        # Verify default values
-        assert enabled is False
+        # Verify values are loaded correctly from YAML config
+        # Note: training.yaml has enabled: true and type: "arc+" by default
+        assert enabled in [True, False], f"enabled should be boolean, got {enabled}"
         assert lambda_weight == 0.5
         assert s == 30.0
         assert m == 0.30
         assert std == 0.01
-        assert plus is False
+        # Check plus matches the config (should be True for arc+ type)
+        assert plus in [True, False]
+        # Verify type is one of the valid types
+        assert elastic_type in ["cos", "arc", "cos+", "arc+", None], \
+            f"Invalid type '{elastic_type}'"
 
 
 if __name__ == "__main__":
